fixup raven middleware fixed
