+++ package.json
@@ -41,7 +41,7 @@
     "nodemon": "^1.9.1",
     "raf": "^3.2.0",
     "raven": "^1.1.1",
-    "raven-js": "3.7.0",
+    "raven-js": "^3.7.0",
     "react": "15.0.2",
     "react-dom": "15.0.2",
     "react-motion": "0.4.4",
